requirements.txt
pytorch-lightning
hydra-core
transformers
sklearn
lightning-bolts
wandb